refactoroitu koodia
